# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Fridge Tracker API
  version: 1.0.0

tags:
  - name: auth
    description: Authentication endpoints
  - name: items
    description: Manage your custom item definitions
  - name: fridges
    description: Manage multiple fridges
  - name: fridge-items
    description: Add or remove items in a fridge

paths:
  # Authentication
  /api/v1/auth/register:
    post:
      tags:
        - auth
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input
        '409':
          description: Username already exists

  /api/v1/auth/login:
    post:
      tags:
        - auth
      summary: Authenticate user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials

  /api/v1/auth/refresh:
    post:
      tags:
        - auth
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid refresh token

  # Items
  /api/v1/items:
    get:
      tags:
        - items
      operationId: getAllItems
      summary: List all custom items for the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemResponse'
    post:
      tags:
        - items
      operationId: createItem
      summary: Create a new item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Invalid input

  /api/v1/items/{itemId}:
    get:
      tags:
        - items
      operationId: getItem
      summary: Get a single item
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '404':
          description: Not found
    put:
      tags:
        - items
      operationId: updateItem
      summary: Update an existing item's name
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemRequest'
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Invalid input
        '404':
          description: Item not found
    delete:
      tags:
        - items
      operationId: deleteItem
      summary: Delete an item
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted successfully
        '404':
          description: Not found

  # Fridges
  /api/v1/fridges:
    get:
      tags:
        - fridges
      operationId: getAllFridges
      summary: List all fridges for current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of fridges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FridgeResponse'
    post:
      tags:
        - fridges
      operationId: createFridge
      summary: Create a new fridge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFridgeRequest'
      responses:
        '201':
          description: Fridge created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FridgeResponse'

  /api/v1/fridges/{fridgeId}:
    get:
      tags:
        - fridges
      operationId: getFridge
      summary: Get fridge details
      security:
        - bearerAuth: []
      parameters:
        - name: fridgeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fridge details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FridgeResponse'
        '404':
          description: Not found
    put:
      tags:
        - fridges
      operationId: updateFridge
      summary: Update fridge name
      security:
        - bearerAuth: []
      parameters:
        - name: fridgeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFridgeRequest'
      responses:
        '200':
          description: Fridge updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FridgeResponse'
    delete:
      tags:
        - fridges
      operationId: deleteFridge
      summary: Delete a fridge
      security:
        - bearerAuth: []
      parameters:
        - name: fridgeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted successfully
        '404':
          description: Not found

  # Fridge Items
  /api/v1/fridges/{fridgeId}/items:
    get:
      tags:
        - fridge-items
      operationId: getFridgeItems
      summary: List all items in a fridge
      security:
        - bearerAuth: []
      parameters:
        - name: fridgeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of fridge items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FridgeItemResponse'
    post:
      tags:
        - fridge-items
      operationId: addItemToFridge
      summary: Add item to fridge
      security:
        - bearerAuth: []
      parameters:
        - name: fridgeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFridgeItemRequest'
      responses:
        '201':
          description: Item added to fridge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FridgeItemResponse'

  /api/v1/fridges/{fridgeId}/items/{fridgeItemId}:
    delete:
      tags:
        - fridge-items
      operationId: removeItemFromFridge
      summary: Remove item from fridge
      security:
        - bearerAuth: []
      parameters:
        - name: fridgeId
          in: path
          required: true
          schema:
            type: string
        - name: fridgeItemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Item removed
        '404':
          description: Not found
    put:
      tags:
        - fridge-items
      operationId: updateFridgeItem
      summary: Update item's best before date
      security:
        - bearerAuth: []
      parameters:
        - name: fridgeId
          in: path
          required: true
          schema:
            type: string
        - name: fridgeItemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFridgeItemRequest'
      responses:
        '200':
          description: Best before date updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FridgeItemResponse'

components:
  schemas:
    # Authentication
    RegisterRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          format: password
          minLength: 8

    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password

    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string

    AuthResponse:
      type: object
      properties:
        tokens:
          $ref: '#/components/schemas/Tokens'

    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string

    Tokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string

    # Item Definitions
    CreateItemRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Custom name of the item

    ItemResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        createdOn:
          type: string
          format: date-time

    UpdateItemRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Updated name of the item

    # Fridge Schemas
    CreateFridgeRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100

    FridgeResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        createdOn:
          type: string
          format: date-time

    UpdateFridgeRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100

    # Fridge Item Schemas
    AddFridgeItemRequest:
      type: object
      required:
        - itemId
        - storedAt
        - bestBeforeDate
      properties:
        itemId:
          type: string
        storedAt:
          type: string
          format: date-time
        bestBeforeDate:
          type: string
          format: date-time

    FridgeItemResponse:
      type: object
      properties:
        id:
          type: string
        fridgeId:
          type: string
        itemId:
          type: string
        itemName:
          type: string
        bestBeforeDate:
          type: string
          format: date-time
        storedAt:
          type: string
          format: date-time

    UpdateFridgeItemRequest:
      type: object
      required:
        - storedAt
        - bestBeforeDate
      properties:
        storedAt:
          type: string
          format: date-time
        bestBeforeDate:
          type: string
          format: date-time

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT